sprint3/app/requirements.txt
Django==4.2.4
Pillow==9.5.0
django-browser-reload==1.8.0
django-widget-tweaks==1.4.12
wheel==0.40.0
django-environ==0.10.0
django-ckeditor==6.5.1
aiosmtpd==1.4.2
argon2-cffi==23.1.0